Fixed the prototype for generate_2d_array to allow int or long dimension sizes.
git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@11996 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/lib/nfp/wrapper.c
@@ -7149,7 +7149,7 @@ void NclAddUserFuncs(void)
     SetArgTemplate(args, nargs, "numeric", 1, dimsizes);  nargs++;
     SetArgTemplate(args, nargs, "integer", 1, dimsizes);  nargs++;
     dimsizes[0] = 2;
-    SetArgTemplate(args, nargs, "integer", 1, dimsizes);  nargs++;
+    SetArgTemplate(args, nargs, "numeric", 1, dimsizes);  nargs++;
 
     NclRegisterFunc(generate_2d_array_W, args, "generate_2d_array", nargs);
